requirements.txt
garminconnect
python-dotenv
google-genai
python-telegram-bot